Typo (#1)
Starting my morning with espresso and some template literals, and I noticed a super trivial typo. But multiplied by the number of readers, I figure you'll want to fix it. (Love the Magic reference!)

In any case, I'm a huge fan of your writing and wanted to say thanks for this and all the other fantastic posts I've enjoyed. I've got your book on my to-read list. Thanks again!